project cuoi ky springboot

+++ mobile_stores/src/main/java/com/example/mobile_stores/services/ProductsRepository.java
@@ -3,7 +3,6 @@ package com.example.mobile_stores.services;
 import com.example.mobile_stores.models.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductsRepository extends JpaRepository<Product, Integer> {
-
+public interface ProductsRepository extends JpaRepository<Product, Long> {
 
 }
